#917: PGS Deliverable Accomplisment  - Added RemoveAttachment property to PgsDeliverableAccomplishmentForm  - Revised MapPut method to implement RemoveAttachment

+++ backend/IMIS/IMIS.Presentation/PgsDeliverableAccomplishmentModule/PgsDeliverableAccomplishmentEndPoints.cs
@@ -126,20 +126,29 @@ namespace IMIS.Presentation.PgsDeliverableAccomplishmentModule
                 }
                 else
                 {
-                    if (!string.IsNullOrWhiteSpace(existing.AttachmentPath))
+                   
+                    if (form.RemoveAttachment)
                     {
-                        var fullPath = existing.AttachmentPath.Replace("\\", "/");
-
-                        var directory = Path.GetDirectoryName(fullPath)?.Replace("\\", "/");
-                        var fileName = Path.GetFileName(fullPath);
-
-                        if (!string.IsNullOrWhiteSpace(directory) && !string.IsNullOrWhiteSpace(fileName))
+                        if (!string.IsNullOrWhiteSpace(existing.AttachmentPath))
                         {
-                            await FTPHelper.DeleteAsync(directory, fileName, cancellationToken);
+                            var fullPath = existing.AttachmentPath.Replace("\\", "/");
+
+                            var directory = Path.GetDirectoryName(fullPath)?.Replace("\\", "/");
+                            var fileName = Path.GetFileName(fullPath);
+
+                            if (!string.IsNullOrWhiteSpace(directory) && !string.IsNullOrWhiteSpace(fileName))
+                            {
+                                await FTPHelper.DeleteAsync(directory, fileName, cancellationToken);
+                            }
                         }
+
+                        uploadedFilePath = null;
+                    }
+                    else
+                    {
+                        uploadedFilePath = existing.AttachmentPath;
                     }
 
-                    uploadedFilePath = null;
                 }
 
                 var dto = new PgsDeliverableAccomplishmentDto
